requirements.txt
discord.py>=2.3
aiohttp
aiosqlite
python-dotenv
APScheduler
pydantic-settings>=2
